Remove markdown and rename MainPage
