--- src/main/resources/welcome.md
@@ -1,3 +1,0 @@
-# Hellow world
-
-I am a _cool_ markdown file in `src/main/resources/`
